moved ManifesMetadataList to Manifest store from api utils
